refactor(js) : convert single-line if blocks into guard clauses

+++ wardrobe/src/app/clothes/detail/[id]/sections/hard_delete_clothes_used_by_id.js
@@ -27,9 +27,7 @@ export default function HardDeleteClothesUsedById(props) {
                 confirmButtonText: "Yes, Delete it!",
                 cancelButtonText: "No, Cancel!"
             }).then(async (result) => {
-                if (result.isConfirmed) {
-                    deleteClothesUsedById(id,tokenKey,props.fetchClothes())
-                } 
+                if (result.isConfirmed) deleteClothesUsedById(id,tokenKey,props.fetchClothes())
             })
         }
     }
